+++ Source/Internal/IGListSectionControllerInternal.h
@@ -19,4 +19,8 @@ FOUNDATION_EXTERN void IGListSectionControllerPopThread(void);
 
 @property (nonatomic, weak, readwrite) UIViewController *viewController;
 
+@property (nonatomic, assign, readwrite) BOOL isFirstSection;
+
+@property (nonatomic, assign, readwrite) BOOL isLastSection;
+
 @end